+++ examples/Searchspaces/continuous_space_botorch_function.py
@@ -62,7 +62,7 @@ parameters = [
 # two-dimensional tuples.
 
 # Construct searchspace, objective and BayBE object.
-searchspace = SearchSpace.create(parameters=parameters)
+searchspace = SearchSpace.from_product(parameters=parameters)
 
 objective = Objective(
     mode="SINGLE", targets=[NumericalTarget(name="Target", mode="MIN")]